[TASK] Disentangle SystemFieldPalettesInterface from SimpleTcaSchema
